Moved the chart to a different folder
