+++ FrontEnd/src/pages/AdminPage/Charts/ActivitiesBarChart.jsx
