+++ FrontEnd/src/pages/AdminPage/Charts/ActivitiesBarChart.module.css
